move auth to composable, add signOut
