+++ frontend/graphql/revokeSession.gql
@@ -1,0 +1,5 @@
+mutation revokeSession($sessionId: String!) {
+    revokeSession(sessionId: $sessionId) {
+        id
+    }
+}
