+++ frontend/package.json
@@ -22,6 +22,7 @@
     "@huntersofbook/naive-ui-nuxt": "^0.5.1",
     "@nuxtjs/eslint-config-typescript": "^12.0.0",
     "@nuxtjs/tailwindcss": "^6.4.0",
+    "@types/jsonwebtoken": "^9.0.1",
     "@types/node": "^18.11.18",
     "@vicons/fluent": "^0.12.0",
     "@vicons/ionicons5": "^0.12.0",
@@ -32,6 +33,7 @@
   },
   "dependencies": {
     "@vueuse/core": "^9.13.0",
+    "jwt-decode": "^3.1.2",
     "nuxt-graphql-client": "^0.2.27"
   }
 }
